refactor: 주문 도메인 분리, User 도메인 내부 feature 단위 리졸버, SDL 분리
