+++ src/features/seller/services/seller-base.service.ts
@@ -113,48 +113,6 @@ export abstract class SellerBaseService {
     }
   }
 
-  protected assertOrderStatusTransition(
-    from: OrderStatus,
-    to: OrderStatus,
-  ): void {
-    if (from === to) {
-      throw new BadRequestException('Order status is already set to target.');
-    }
-
-    if (to === OrderStatus.CONFIRMED && from !== OrderStatus.SUBMITTED) {
-      throw new BadRequestException('Invalid order status transition.');
-    }
-
-    if (to === OrderStatus.MADE && from !== OrderStatus.CONFIRMED) {
-      throw new BadRequestException('Invalid order status transition.');
-    }
-
-    if (to === OrderStatus.PICKED_UP && from !== OrderStatus.MADE) {
-      throw new BadRequestException('Invalid order status transition.');
-    }
-
-    if (to === OrderStatus.CANCELED) {
-      const cancellable =
-        from === OrderStatus.SUBMITTED ||
-        from === OrderStatus.CONFIRMED ||
-        from === OrderStatus.MADE;
-      if (!cancellable) {
-        throw new BadRequestException(
-          'Order cannot be canceled from current status.',
-        );
-      }
-    }
-  }
-
-  protected toOrderStatus(raw: string): OrderStatus {
-    if (raw === 'SUBMITTED') return OrderStatus.SUBMITTED;
-    if (raw === 'CONFIRMED') return OrderStatus.CONFIRMED;
-    if (raw === 'MADE') return OrderStatus.MADE;
-    if (raw === 'PICKED_UP') return OrderStatus.PICKED_UP;
-    if (raw === 'CANCELED') return OrderStatus.CANCELED;
-    throw new BadRequestException('Invalid order status.');
-  }
-
   protected toConversationBodyFormat(raw: string): ConversationBodyFormat {
     if (raw === 'TEXT') return ConversationBodyFormat.TEXT;
     if (raw === 'HTML') return ConversationBodyFormat.HTML;
