+++ src/features/user/user-notification.graphql
@@ -1,0 +1,59 @@
+extend type Query {
+  """상단/하단 공통 UI용 카운트 정보 조회"""
+  viewerCounts: ViewerCounts!
+  """알림 목록 조회"""
+  myNotifications(input: MyNotificationsInput): NotificationConnection!
+}
+
+extend type Mutation {
+  """특정 알림 읽음 처리"""
+  markNotificationRead(notificationId: ID!): Boolean!
+  """전체 알림 읽음 처리"""
+  markAllNotificationsRead: Boolean!
+}
+
+"""뷰어 카운트 정보"""
+type ViewerCounts {
+  """읽지 않은 알림 수"""
+  unreadNotificationCount: Int!
+  """장바구니 아이템 수"""
+  cartItemCount: Int!
+  """위시리스트 아이템 수"""
+  wishlistCount: Int!
+}
+
+"""알림 목록 조회 입력"""
+input MyNotificationsInput {
+  """읽지 않은 알림만 조회"""
+  unreadOnly: Boolean = false
+  """오프셋"""
+  offset: Int = 0
+  """조회 개수(최대 50)"""
+  limit: Int = 20
+}
+
+"""알림 목록 응답"""
+type NotificationConnection {
+  """알림 아이템 목록"""
+  items: [NotificationItem!]!
+  """전체 개수"""
+  totalCount: Int!
+  """다음 페이지 존재 여부"""
+  hasMore: Boolean!
+}
+
+"""알림 아이템"""
+type NotificationItem {
+  """알림 ID"""
+  id: ID!
+  """알림 타입"""
+  type: NotificationType!
+  """제목"""
+  title: String!
+  """내용"""
+  body: String!
+  """읽음 처리 시각"""
+  readAt: DateTime
+  """생성 시각"""
+  createdAt: DateTime!
+}
