+++ src/features/user/user-search.graphql
@@ -1,0 +1,39 @@
+extend type Query {
+  """최근 검색어 목록 조회"""
+  mySearchHistories(input: MySearchHistoriesInput): SearchHistoryConnection!
+}
+
+extend type Mutation {
+  """특정 검색어 기록 삭제"""
+  deleteSearchHistory(id: ID!): Boolean!
+  """검색어 기록 전체 삭제"""
+  clearSearchHistories: Boolean!
+}
+
+"""검색어 목록 조회 입력"""
+input MySearchHistoriesInput {
+  """오프셋"""
+  offset: Int = 0
+  """조회 개수(최대 50)"""
+  limit: Int = 20
+}
+
+"""검색어 목록 응답"""
+type SearchHistoryConnection {
+  """검색어 아이템 목록"""
+  items: [SearchHistoryItem!]!
+  """전체 개수"""
+  totalCount: Int!
+  """다음 페이지 존재 여부"""
+  hasMore: Boolean!
+}
+
+"""검색어 아이템"""
+type SearchHistoryItem {
+  """검색 기록 ID"""
+  id: ID!
+  """검색어"""
+  keyword: String!
+  """마지막 사용 시각"""
+  lastUsedAt: DateTime!
+}
