+++ src/features/user/user.module.ts
@@ -1,8 +1,13 @@
 import { Module } from '@nestjs/common';
 
 import { UserRepository } from './repositories/user.repository';
-import { UserMutationResolver } from './resolvers/user-mutation.resolver';
-import { UserQueryResolver } from './resolvers/user-query.resolver';
+import { UserEngagementMutationResolver } from './resolvers/user-engagement-mutation.resolver';
+import { UserNotificationMutationResolver } from './resolvers/user-notification-mutation.resolver';
+import { UserNotificationQueryResolver } from './resolvers/user-notification-query.resolver';
+import { UserProfileMutationResolver } from './resolvers/user-profile-mutation.resolver';
+import { UserProfileQueryResolver } from './resolvers/user-profile-query.resolver';
+import { UserSearchMutationResolver } from './resolvers/user-search-mutation.resolver';
+import { UserSearchQueryResolver } from './resolvers/user-search-query.resolver';
 import { UserService } from './user.service';
 
 /**
@@ -12,8 +17,13 @@ import { UserService } from './user.service';
   providers: [
     UserService,
     UserRepository,
-    UserQueryResolver,
-    UserMutationResolver,
+    UserProfileQueryResolver,
+    UserNotificationQueryResolver,
+    UserSearchQueryResolver,
+    UserProfileMutationResolver,
+    UserNotificationMutationResolver,
+    UserSearchMutationResolver,
+    UserEngagementMutationResolver,
   ],
 })
 export class UserModule {}
